requirements.txt
asgiref==3.7.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.2
django-auditlog==2.3.0
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-push-notifications==3.0.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-tenant-schemas==1.11.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.5
ecdsa==0.18.0
google-api-core==2.11.1
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jwcrypto==1.5.0
Markdown==3.4.3
MarkupSafe==2.1.3
newrelic==9.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
ordered-set==4.1.0
packaging==23.2
Pillow==10.1.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-jose==3.3.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
